(chore) move request-client helper file to src/helpers -- accidentally put it in newList helpers
