+++ frontend/src/gql/hooks/list/useCreateList.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { gql } from "graphql-request";
-import requestClient from "../../../components/newlist/helpers/request-client";
+import requestClient from "../../../helpers/request-client";
 import { List, NewListWithTermsInput } from "../../codegen-output";
 import { listPropsFragment } from "../../fragments/list-fragments";
 import { termPropsFragment } from "../../fragments/term-fragments";
