+++ frontend/src/gql/hooks/user/useLogin.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { gql } from "graphql-request";
-import requestClient from "../../../components/newlist/helpers/request-client";
+import requestClient from "../../../helpers/request-client";
 import { User, UserInput } from "../../codegen-output";
 
 export const loginMutation = gql`
