+++ frontend/src/gql/hooks/user/useLogout.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { gql } from "graphql-request";
-import requestClient from "../../../components/newlist/helpers/request-client";
+import requestClient from "../../../helpers/request-client";
 import { Message } from "../../codegen-output";
 
 const logoutMutation = gql`
