+++ frontend/src/gql/hooks/user/useQueryMe.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { gql } from "graphql-request";
-import requestClient from "../../../components/newlist/helpers/request-client";
+import requestClient from "../../../helpers/request-client";
 import { User } from "../../codegen-output";
 import { userPropsFragment } from "../../fragments/user-fragments";
